chore: projectResponseDto 이너 클래스로 변경 및 imagePath 오류 수정
