+++ src/main/java/ssuPlector/dto/responseDto/ProjectListResponseDto.java
@@ -6,6 +6,7 @@ import java.util.stream.Collectors;
 import org.springframework.data.domain.Page;
 
 import lombok.Getter;
+import ssuPlector.domain.Image;
 import ssuPlector.domain.Project;
 
 @Getter
@@ -28,5 +29,25 @@ public class ProjectListResponseDto {
         this.totalPage = projectPage.getTotalPages();
         this.currentElement = projectPage.getNumberOfElements();
         this.totalElement = projectPage.getTotalElements();
+    }
+
+    @Getter
+    public class ProjectResponseDto {
+        private Long id;
+        private String name;
+        private String imagePath;
+        private String shortIntro;
+        private String category;
+        private long hits;
+
+        public ProjectResponseDto(Project project) {
+            this.id = project.getId();
+            this.name = project.getName();
+            for (Image image : project.getImageList())
+                if (image.isMainImage()) this.imagePath = image.getImagePath();
+            this.shortIntro = project.getShortIntro();
+            this.category = project.getCategory().name();
+            this.hits = project.getHits();
+        }
     }
 }
